# api/handler/proposal.go
// Copyright 2025 Boyuan-IT-Club
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package handler

import (
	"github.com/Boyuan-IT-Club/Meowpick-Backend/application/dto"
	"github.com/Boyuan-IT-Club/Meowpick-Backend/infra/util/token"
	"github.com/Boyuan-IT-Club/Meowpick-Backend/provider"
	"github.com/Boyuan-IT-Club/Meowpick-Backend/types/consts"
	"github.com/gin-gonic/gin"
)

// CreateProposal godoc
// @Summary 新增提案
// @Description 创建一个新的提案
// @Tags proposal
// @Accept json
// @Param req body dto.CreateProposalReq true "创建提案的请求参数"
// @success 200 {object} Response[dto.CreateProposalResp]
// @Security Bearer
// @Router /api/proposal/add [post]
func CreateProposal(c *gin.Context) {
	var req dto.CreateProposalReq
	var resp *dto.CreateProposalResp
	var err error

	if err = c.ShouldBindJSON(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.CreateProposal(c, &req)
	PostProcess(c, &req, resp, err)
}

// ListProposals godoc
// @Summary 分页获取提案列表
// @Description 分页查询提案列表数据
// @Tags proposal
// @Produce json
// @Param page query int true "页码"
// @Param pageSize query int true "每页数量"
// @Success 200 {object} Response[dto.ListProposalResp]
// @Security Bearer
// @Router /api/proposal/list [get]
func ListProposals(c *gin.Context) {
	var req dto.ListProposalReq
	var resp *dto.ListProposalResp
	var err error

	if err = c.ShouldBindQuery(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.ListProposals(c, &req)
	PostProcess(c, &req, resp, err)
}

// FilterProposals godoc
// @Summary 分页筛选提案列表
// @Description 基于提案状态、校区、开课院系、课程分类筛选 proposal 表中的提案
// @Tags proposal
// @Produce json
// @Param status query []string true "提案状态，可多选" collectionFormat(multi)
// @Param campus query []string true "校区，可多选" collectionFormat(multi)
// @Param department query string false "开课院系，精确匹配"
// @Param category query string false "课程分类，精确匹配"
// @Param page query int false "页码" default(1)
// @Param pageSize query int false "每页数量" default(10)
// @Success 200 {object} Response[dto.ListProposalResp]
// @Security Bearer
// @Router /api/proposal/filter [get]
func FilterProposals(c *gin.Context) {
	var req dto.FilterProposalReq
	var resp *dto.ListProposalResp
	var err error

	if err = c.ShouldBindQuery(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.FilterProposals(c, &req)
	PostProcess(c, &req, resp, err)
}

// GetProposal 获取提案详情
// @Summary 获取提案详情
// @Description 根据提案ID查询提案完整信息
// @Tags proposal
// @Produce json
// @Param id path string true "提案ID"
// @Success 200 {object} Response[dto.GetProposalResp]
// @Security Bearer
// @Router /api/proposal/{proposalId} [get]
func GetProposal(c *gin.Context) {
	var req dto.GetProposalReq
	var resp *dto.GetProposalResp
	var err error

	req.ProposalID = c.Param(consts.CtxProposalID)
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.GetProposal(c, &req)
	PostProcess(c, &req, resp, err)
}

// ApproveProposal godoc
// @Summary 审批提案
// @Description 通过提案并创建课程，或拒绝提案
// @Tags proposal
// @Produce json
// @Param proposalId path string true "提案ID"
// @Success 200 {object} Response[dto.ToggleProposalResp]
// @Security Bearer
// @Router /api/proposal/{proposalId}/approve [post]
func ApproveProposal(c *gin.Context) {
	var req dto.ToggleProposalReq
	var resp *dto.ToggleProposalResp
	var err error

	req.ProposalID = c.Param(consts.CtxProposalID)
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.ApproveProposal(c, &req)
	PostProcess(c, &req, resp, err)
}

// RevokeProposal godoc
// @Summary 撤回提案操作
// @Description 管理员撤回提案的通过/拒绝/删除操作
// @Tags proposal
// @Accept json
// @Param proposalId path string true "提案ID"
// @Param req body dto.RevokeProposalReq true "撤回操作类型"
// @Success 200 {object} Response[dto.RevokeProposalResp]
// @Security Bearer
// @Router /api/proposal/{proposalId}/revoke [post]
func RevokeProposal(c *gin.Context) {
	var req dto.RevokeProposalReq
	var resp *dto.RevokeProposalResp
	var err error

	req.ProposalID = c.Param(consts.CtxProposalID)

	if err = c.ShouldBindJSON(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.RevokeProposal(c, &req)
	PostProcess(c, &req, resp, err)
}

// RejectProposal godoc
// @Summary 拒绝提案
// @Description 管理员操作：将状态为 pending（待审核）的提案变更为 rejected（已拒绝）
// @Description 使用场景：课程提案审核流程中，管理员认为提案不符合要求，驳回该提案
// @Description 注意事项：
// @Description - 仅管理员可操作（需先调用 /api/auth/is_admin 确认权限）
// @Description - 仅状态为 pending 的提案可以拒绝，已 approved/rejected 的提案无法再次操作
// @Description - 拒绝后不会创建课程记录，仅更新提案状态
// @Tags proposal
// @Accept json
// @Produce json
// @Param proposalId path string true "提案ID"
// @Param body body dto.RejectProposalReq true "拒绝参数（可选理由）"
// @Success 200 {object} Response[dto.RejectProposalResp]
// @Security Bearer
// @Router /api/proposal/{proposalId}/reject [post]
func RejectProposal(c *gin.Context) {
	var req dto.RejectProposalReq
	var resp *dto.RejectProposalResp
	var err error

	if err = c.ShouldBindJSON(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}
	req.ProposalID = c.Param(consts.CtxProposalID)
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.RejectProposal(c, &req)
	PostProcess(c, &req, resp, err)
}

// UpdateProposal 更新提案接口
// @Summary 更新提案内容
// @Description 根据提案ID修改提案的标题和内容
// @Tags proposal
// @Accept json
// @Produce json
// @Param proposalId path string true "提案唯一ID"
// @Param body body dto.UpdateProposalReq true "更新参数（标题、内容）"
// @Success 200 {object} Response[dto.UpdateProposalResp] "更新成功响应"
// @Security Bearer
// @Router /api/proposal/{proposalId}/update [post]
func UpdateProposal(c *gin.Context) {
	var req dto.UpdateProposalReq
	var resp *dto.UpdateProposalResp
	var err error

	req.ProposalID = c.Param(consts.CtxProposalID)

	if err = c.ShouldBindJSON(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}

	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.UpdateProposal(c, &req)

	PostProcess(c, &req, resp, err)
}

// DeleteProposal godoc
// @Summary 删除提案
// @Description 根据提案ID软删除提案（标记为已删除状态）,只对状态为pending和rejected的进行处理
// @Tags proposal
// @Accept json
// @Param proposalId path string true "提案ID"
// @success 200 {object} Response[dto.DeleteProposalResp]
// @Security Bearer
// @Router /api/proposal/{proposalId}/delete [POST]
func DeleteProposal(c *gin.Context) {
	var err error
	var req dto.DeleteProposalReq
	var resp *dto.DeleteProposalResp

	req.ProposalID = c.Param(consts.CtxProposalID)

	if err = c.ShouldBindJSON(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.DeleteProposal(c, &req)
	PostProcess(c, &req, resp, err)
}

// GetProposalSuggestions godoc
// @Summary 获取提案搜索建议
// @Description 根据关键词模糊分页搜索提案标题，返回匹配的提案建议列表
// @Tags proposal
// @Produce json
// @Param keyword query string true "搜索关键词"
// @Param page query int false "页码" default(0)
// @Param pageSize query int false "每页数量" default(10)
// @Success 200 {object} Response[dto.GetProposalSuggestionsResp]
// @Security Bearer
// @Router /api/proposal/suggest [post]
func GetProposalSuggestions(c *gin.Context) {
	var req dto.GetProposalSuggestionsReq
	var resp *dto.GetProposalSuggestionsResp
	var err error

	if err = c.ShouldBindQuery(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.GetProposalSuggestions(c, &req)
	PostProcess(c, &req, resp, err)
}

// GetProposalFieldSuggestions godoc
// @Summary 获取提案字段建议
// @Description 根据字段类型和关键词获取建议列表，支持学院、类别、校区、课程名称、课程代码、教师姓名
// @Tags proposal
// @Produce json
// @Param field query string true "字段类型: department/category/campus/courseName/courseCode/teacherName"
// @Param keyword query string true "搜索关键词"
// @Param page query int false "页码" default(0)
// @Param pageSize query int false "每页数量" default(10)
// @Success 200 {object} dto.GetProposalFieldSuggestionsResp
// @Security Bearer
// @Router /api/proposal/field-suggestions [get]
func GetProposalFieldSuggestions(c *gin.Context) {
	var req dto.GetProposalFieldSuggestionsReq
	var resp *dto.GetProposalFieldSuggestionsResp
	var err error

	if err = c.ShouldBindQuery(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.GetProposalFieldSuggestions(c, &req)
	PostProcess(c, &req, resp, err)
}

// GetMyProposals godoc
// @Summary 获取我的提案
// @Tags proposal
// @Param page query int false "页码" default(1)
// @Param pageSize query int false "每页数量" default(10)
// @Param status query string false "状态" Enums(pending,approved,rejected)
// @Success 200 {object} Response[dto.GetMyProposalsResp]
// @Security Bearer
// @Router /api/proposal/history [get]
func GetMyProposals(c *gin.Context) {
	var req dto.GetMyProposalsReq
	var resp *dto.GetMyProposalsResp
	var err error

	if err = c.ShouldBindQuery(&req); err != nil {
		PostProcess(c, &req, nil, err)
		return
	}
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().ProposalService.GetMyProposals(c, &req)
	PostProcess(c, &req, resp, err)
}