# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "dto.CreateProposalReq": {
                "properties": {
                    "content": {
                        "type": "string"
                    },
                    "course": {
                        "$ref": "#/components/schemas/dto.ProposalCourseVO"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "required": [
                    "content",
                    "course",
                    "title"
                ],
                "type": "object"
            },
...
            "dto.GetProposalFieldSuggestionsResp": {
                "properties": {
                    "field": {
                        "type": "string"
                    },
                    "suggestions": {
                        "items": {
                            "$ref": "#/components/schemas/dto.FieldSuggestionVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_CreateProposalResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.CreateProposalResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_ListProposalResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.ListProposalResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "选课猫后端接口文档",
        "title": "Meowpick Backend API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/proposal/add": {
            "post": {
                "description": "创建一个新的提案",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.CreateProposalReq",
                                        "summary": "req",
                                        "description": "创建提案的请求参数"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "创建提案的请求参数",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_CreateProposalResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "新增提案",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/field-suggestions": {
            "get": {
                "description": "根据字段类型和关键词获取建议列表，支持学院、类别、校区、课程名称、课程代码、教师姓名",
                "parameters": [
                    {
                        "description": "字段类型: department/category/campus/courseName/courseCode/teacherName",
                        "in": "query",
                        "name": "field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "搜索关键词",
                        "in": "query",
                        "name": "keyword",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "页码",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "每页数量",
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "default": 10,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.GetProposalFieldSuggestionsResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取提案字段建议",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/filter": {
            "get": {
                "description": "基于提案状态、校区、开课院系、课程分类筛选 proposal 表中的提案",
                "parameters": [
                    {
                        "description": "提案状态，可多选",
                        "in": "query",
                        "name": "status",
                        "required": true,
                        "schema": {
                            "items": {
                                "type": "string"
                            },
                            "type": "array"
                        },
                        "style": "form"
                    },
                    {
                        "description": "校区，可多选",
                        "in": "query",
                        "name": "campus",
                        "required": true,
                        "schema": {
                            "items": {
                                "type": "string"
                            },
                            "type": "array"
                        },
                        "style": "form"
                    },
                    {
                        "description": "开课院系，精确匹配",
                        "in": "query",
                        "name": "department",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "课程分类，精确匹配",
                        "in": "query",
                        "name": "category",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "页码",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "每页数量",
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "default": 10,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_ListProposalResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "分页筛选提案列表",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/history": {
            "get": {
                "parameters": [
                    {
                        "description": "页码",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "每页数量",
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "default": 10,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "状态",
                        "in": "query",
                        "name": "status",
                        "schema": {
                            "enum": [
                                "pending",
                                "approved",
                                "rejected"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetMyProposalsResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取我的提案",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/list": {
            "get": {
                "description": "分页查询提案列表数据",
                "parameters": [
                    {
                        "description": "页码",
                        "in": "query",
                        "name": "page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "每页数量",
                        "in": "query",
                        "name": "pageSize",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_ListProposalResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "分页获取提案列表",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/suggest": {
            "post": {
                "description": "根据关键词模糊分页搜索提案标题，返回匹配的提案建议列表",
                "parameters": [
                    {
                        "description": "搜索关键词",
                        "in": "query",
                        "name": "keyword",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "页码",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "每页数量",
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "default": 10,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetProposalSuggestionsResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取提案搜索建议",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/{proposalId}": {
            "get": {
                "description": "根据提案ID查询提案完整信息",
                "parameters": [
                    {
                        "description": "提案ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetProposalResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取提案详情",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/{proposalId}/approve": {
            "post": {
                "description": "通过提案并创建课程，或拒绝提案",
                "parameters": [
                    {
                        "description": "提案ID",
                        "in": "path",
                        "name": "proposalId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_ToggleProposalResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "审批提案",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/{proposalId}/delete": {
            "post": {
                "description": "根据提案ID软删除提案（标记为已删除状态）,只对状态为pending和rejected的进行处理",
                "parameters": [
                    {
                        "description": "提案ID",
                        "in": "path",
                        "name": "proposalId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_DeleteProposalResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "删除提案",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/{proposalId}/reject": {
            "post": {
                "description": "管理员操作：将状态为 pending（待审核）的提案变更为 rejected（已拒绝）\n使用场景：课程提案审核流程中，管理员认为提案不符合要求，驳回该提案\n注意事项：\n- 仅管理员可操作（需先调用 /api/auth/is_admin 确认权限）\n- 仅状态为 pending 的提案可以拒绝，已 approved/rejected 的提案无法再次操作\n- 拒绝后不会创建课程记录，仅更新提案状态",
                "parameters": [
                    {
                        "description": "提案ID",
                        "in": "path",
                        "name": "proposalId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.RejectProposalReq",
                                        "summary": "body",
                                        "description": "拒绝参数（可选理由）"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "拒绝参数（可选理由）",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_RejectProposalResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "拒绝提案",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/{proposalId}/revoke": {
            "post": {
                "description": "管理员撤回提案的通过/拒绝/删除操作",
                "parameters": [
                    {
                        "description": "提案ID",
                        "in": "path",
                        "name": "proposalId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.RevokeProposalReq",
                                        "summary": "req",
                                        "description": "撤回操作类型"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "撤回操作类型",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_RevokeProposalResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "撤回提案操作",
                "tags": [
                    "proposal"
                ]
            }
        },
        "/api/proposal/{proposalId}/update": {
            "post": {
                "description": "根据提案ID修改提案的标题和内容",
                "parameters": [
                    {
                        "description": "提案唯一ID",
                        "in": "path",
                        "name": "proposalId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.UpdateProposalReq",
                                        "summary": "body",
                                        "description": "更新参数（标题、内容）"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "更新参数（标题、内容）",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_UpdateProposalResp"
                                }
                            }
                        },
                        "description": "更新成功响应"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "更新提案内容",
                "tags": [
                    "proposal"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}